requirements.txt
fastapi==0.110.1
Jinja2==3.1.3
pydantic==2.6.2
uvicorn==0.29.0
python-dotenv==1.0.1
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
supabase==2.4.5